requirements.txt
PyJWT==2.6.0
pytz==2022.7.1

pip~=21.2.4
wheel~=0.37.0
setuptools~=58.0.4
future~=0.18.2
requests~=2.28.1
fastapi~=0.82.0
python-dotenv~=0.21.0
SQLAlchemy~=1.4.41
PyMySQL~=1.0.2